feat: extract config and ui modules
